feat: Activity links
I found it quite a bit distracting that I could not open activity detail
in a new window - especially as filters on the matrix page do not
persist on navigation.

This changes the detail navigation from an onclick to a regular ol' HTML
link, allowing "open in new window" etc.