Add Level of an Activity to Excel
Since we want to use the Excel file as a reference to map against
a companies ISO requirements, we need also the level attribute.
Because we want to conenctrate on lower level.

This adds the levle property from the parsed YAML object to the temp
object and to the allMappingDataSortedByISO17 which seem to be the
source for the hiden Excel table.

But level column is still empty in resulting Excel file.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>